requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.74.1
Flask==2.0.3
fonttools==4.29.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
keyboard==0.13.5
kiwisolver==1.3.2
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib==3.5.1
MouseInfo==0.1.3
numpy==1.22.2
packaging==21.3
passlib==1.7.4
Pillow==9.0.1
prettify==0.1.1
psycopg2==2.9.3
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.9.0
PyGetWindow==0.0.9
pymongo==4.0.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytweening==1.0.4
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
Werkzeug==2.0.3
